+++ Telemed/DTOs/PaymentDto/CreatePaymentDto.cs
@@ -4,8 +4,8 @@ namespace Telemed.DTOs;
 public class CreatePaymentDto
 {
     public int Appointmentid { get; set; }
-    public int Patientid { get; set; }
-    public int Providerid { get; set; }
+    public long Patientid { get; set; }
+    public long Providerid { get; set; }
     public decimal Amount { get; set; }
     public string? Paymentmethod { get; set; }
 }
