+++ Telemed/DTOs/PaymentDto/PaymentResponseDto.cs
@@ -3,21 +3,21 @@ namespace Telemed.DTOs;
 
 public class PaymentResponseDto
 {
-    public int Paymentid { get; set; }
+    public long Paymentid { get; set; }
 
     // Appointment Info
-    public int Appointmentid { get; set; }
+    public long Appointmentid { get; set; }
     public DateTime? Appointmentdate { get; set; }
     public string? Appointmentmode { get; set; }
 
     // Patient Info
-    public int Patientid { get; set; }
+    public long Patientid { get; set; }
     public string? Patientname { get; set; }
     public string? Patientemail { get; set; }
     public string? Patientphone { get; set; }
 
     // Provider Info
-    public int Providerid { get; set; }
+    public long Providerid { get; set; }
     public string? Providername { get; set; }
     public string? Speciality { get; set; }
