+++ Telemed/DTOs/PrescriptionDto/CreatePrescriptionDto.cs
@@ -4,8 +4,8 @@ namespace Telemed.DTOs;
 public class CreatePrescriptionDto
 {
     public int Encounterid { get; set; }
-    public int Patientid { get; set; }
-    public int Providerid { get; set; }
+    public long Patientid { get; set; }
+    public long Providerid { get; set; }
     public string Medicinename { get; set; } = null!;
     public string? Dosage { get; set; }
     public string? Frequency { get; set; }
