+++ Telemed/DTOs/PrescriptionDto/PrescriptionResponseDto.cs
@@ -3,21 +3,21 @@ namespace Telemed.DTOs;
 
 public class PrescriptionResponseDto
 {
-    public int Prescriptionid { get; set; }
+    public long Prescriptionid { get; set; }
 
     // Encounter Info
-    public int Encounterid { get; set; }
+    public long Encounterid { get; set; }
     public string? Diagnosis { get; set; }
     public string? Icd10code { get; set; }
 
     // Patient Info
-    public int Patientid { get; set; }
+    public long Patientid { get; set; }
     public string? Patientname { get; set; }
     public DateTime? Dateofbirth { get; set; }
     public string? Gender { get; set; }
 
     // Provider Info
-    public int Providerid { get; set; }
+    public long Providerid { get; set; }
     public string? Providername { get; set; }
     public string? Speciality { get; set; }
