+++ Telemed/Models/Message.cs
@@ -5,16 +5,15 @@ namespace Telemed.Models;
 
 public partial class Message
 {
-    public int Messageid { get; set; }
+    public long Messageid { get; set; }
 
-    public int Patientid { get; set; }
+    public long Patientid { get; set; }
 
-    public int Providerid { get; set; }
+    public long Providerid { get; set; }
 
     public string Sendertype { get; set; } = null!;
 
     public string Messagetext { get; set; } = null!;
-
     public DateTime? Sentat { get; set; }
 
     public bool? Isread { get; set; }
