+++ Telemed/Models/Prescription.cs
@@ -9,9 +9,9 @@ public partial class Prescription
 
     public int Encounterid { get; set; }
 
-    public int Patientid { get; set; }
+    public long Patientid { get; set; }
 
-    public int Providerid { get; set; }
+    public long Providerid { get; set; }
 
     public string Medicinename { get; set; } = null!;
